+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/BillSumRecordInfoV2.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /** BillSumRecordInfoV2 */
@@ -31,57 +32,57 @@ public class BillSumRecordInfoV2 {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "official_amount")
 
-    private Double officialAmount;
+    private BigDecimal officialAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "official_discount_amount")
 
-    private Double officialDiscountAmount;
+    private BigDecimal officialDiscountAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "truncated_amount")
 
-    private Double truncatedAmount;
+    private BigDecimal truncatedAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "consume_amount")
 
-    private Double consumeAmount;
+    private BigDecimal consumeAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "coupon_amount")
 
-    private Double couponAmount;
+    private BigDecimal couponAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "flexipurchase_coupon_amount")
 
-    private Double flexipurchaseCouponAmount;
+    private BigDecimal flexipurchaseCouponAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "stored_value_card_amount")
 
-    private Double storedValueCardAmount;
+    private BigDecimal storedValueCardAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "debt_amount")
 
-    private Double debtAmount;
+    private BigDecimal debtAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "writeoff_amount")
 
-    private Double writeoffAmount;
+    private BigDecimal writeoffAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "cash_amount")
 
-    private Double cashAmount;
+    private BigDecimal cashAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "credit_amount")
 
-    private Double creditAmount;
+    private BigDecimal creditAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -164,7 +165,7 @@ public class BillSumRecordInfoV2 {
         this.chargingMode = chargingMode;
     }
 
-    public BillSumRecordInfoV2 withOfficialAmount(Double officialAmount) {
+    public BillSumRecordInfoV2 withOfficialAmount(BigDecimal officialAmount) {
         this.officialAmount = officialAmount;
         return this;
     }
@@ -172,15 +173,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：官网价| |参数的约束及描述：官网价|
      * 
      * @return officialAmount */
-    public Double getOfficialAmount() {
+    public BigDecimal getOfficialAmount() {
         return officialAmount;
     }
 
-    public void setOfficialAmount(Double officialAmount) {
+    public void setOfficialAmount(BigDecimal officialAmount) {
         this.officialAmount = officialAmount;
     }
 
-    public BillSumRecordInfoV2 withOfficialDiscountAmount(Double officialDiscountAmount) {
+    public BillSumRecordInfoV2 withOfficialDiscountAmount(BigDecimal officialDiscountAmount) {
         this.officialDiscountAmount = officialDiscountAmount;
         return this;
     }
@@ -188,15 +189,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：折扣金额| |参数的约束及描述：折扣金额|
      * 
      * @return officialDiscountAmount */
-    public Double getOfficialDiscountAmount() {
+    public BigDecimal getOfficialDiscountAmount() {
         return officialDiscountAmount;
     }
 
-    public void setOfficialDiscountAmount(Double officialDiscountAmount) {
+    public void setOfficialDiscountAmount(BigDecimal officialDiscountAmount) {
         this.officialDiscountAmount = officialDiscountAmount;
     }
 
-    public BillSumRecordInfoV2 withTruncatedAmount(Double truncatedAmount) {
+    public BillSumRecordInfoV2 withTruncatedAmount(BigDecimal truncatedAmount) {
         this.truncatedAmount = truncatedAmount;
         return this;
     }
@@ -204,15 +205,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：抹零金额| |参数的约束及描述：抹零金额|
      * 
      * @return truncatedAmount */
-    public Double getTruncatedAmount() {
+    public BigDecimal getTruncatedAmount() {
         return truncatedAmount;
     }
 
-    public void setTruncatedAmount(Double truncatedAmount) {
+    public void setTruncatedAmount(BigDecimal truncatedAmount) {
         this.truncatedAmount = truncatedAmount;
     }
 
-    public BillSumRecordInfoV2 withConsumeAmount(Double consumeAmount) {
+    public BillSumRecordInfoV2 withConsumeAmount(BigDecimal consumeAmount) {
         this.consumeAmount = consumeAmount;
         return this;
     }
@@ -220,15 +221,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：应付金额，应付金额 = 官网价-折扣金额-抹零金额| |参数的约束及描述：应付金额，应付金额 = 官网价-折扣金额-抹零金额|
      * 
      * @return consumeAmount */
-    public Double getConsumeAmount() {
+    public BigDecimal getConsumeAmount() {
         return consumeAmount;
     }
 
-    public void setConsumeAmount(Double consumeAmount) {
+    public void setConsumeAmount(BigDecimal consumeAmount) {
         this.consumeAmount = consumeAmount;
     }
 
-    public BillSumRecordInfoV2 withCouponAmount(Double couponAmount) {
+    public BillSumRecordInfoV2 withCouponAmount(BigDecimal couponAmount) {
         this.couponAmount = couponAmount;
         return this;
     }
@@ -236,15 +237,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：代金券金额。| |参数的约束及描述：代金券金额。|
      * 
      * @return couponAmount */
-    public Double getCouponAmount() {
+    public BigDecimal getCouponAmount() {
         return couponAmount;
     }
 
-    public void setCouponAmount(Double couponAmount) {
+    public void setCouponAmount(BigDecimal couponAmount) {
         this.couponAmount = couponAmount;
     }
 
-    public BillSumRecordInfoV2 withFlexipurchaseCouponAmount(Double flexipurchaseCouponAmount) {
+    public BillSumRecordInfoV2 withFlexipurchaseCouponAmount(BigDecimal flexipurchaseCouponAmount) {
         this.flexipurchaseCouponAmount = flexipurchaseCouponAmount;
         return this;
     }
@@ -252,15 +253,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：现金券金额，预留。| |参数的约束及描述：现金券金额，预留。|
      * 
      * @return flexipurchaseCouponAmount */
-    public Double getFlexipurchaseCouponAmount() {
+    public BigDecimal getFlexipurchaseCouponAmount() {
         return flexipurchaseCouponAmount;
     }
 
-    public void setFlexipurchaseCouponAmount(Double flexipurchaseCouponAmount) {
+    public void setFlexipurchaseCouponAmount(BigDecimal flexipurchaseCouponAmount) {
         this.flexipurchaseCouponAmount = flexipurchaseCouponAmount;
     }
 
-    public BillSumRecordInfoV2 withStoredValueCardAmount(Double storedValueCardAmount) {
+    public BillSumRecordInfoV2 withStoredValueCardAmount(BigDecimal storedValueCardAmount) {
         this.storedValueCardAmount = storedValueCardAmount;
         return this;
     }
@@ -268,15 +269,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：储值卡金额，预留。| |参数的约束及描述：储值卡金额，预留。|
      * 
      * @return storedValueCardAmount */
-    public Double getStoredValueCardAmount() {
+    public BigDecimal getStoredValueCardAmount() {
         return storedValueCardAmount;
     }
 
-    public void setStoredValueCardAmount(Double storedValueCardAmount) {
+    public void setStoredValueCardAmount(BigDecimal storedValueCardAmount) {
         this.storedValueCardAmount = storedValueCardAmount;
     }
 
-    public BillSumRecordInfoV2 withDebtAmount(Double debtAmount) {
+    public BillSumRecordInfoV2 withDebtAmount(BigDecimal debtAmount) {
         this.debtAmount = debtAmount;
         return this;
     }
@@ -284,15 +285,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：欠费金额，即从客户账户扣费的时候，客户账户金额不足，欠费的金额。| |参数的约束及描述：欠费金额，即从客户账户扣费的时候，客户账户金额不足，欠费的金额。|
      * 
      * @return debtAmount */
-    public Double getDebtAmount() {
+    public BigDecimal getDebtAmount() {
         return debtAmount;
     }
 
-    public void setDebtAmount(Double debtAmount) {
+    public void setDebtAmount(BigDecimal debtAmount) {
         this.debtAmount = debtAmount;
     }
 
-    public BillSumRecordInfoV2 withWriteoffAmount(Double writeoffAmount) {
+    public BillSumRecordInfoV2 withWriteoffAmount(BigDecimal writeoffAmount) {
         this.writeoffAmount = writeoffAmount;
         return this;
     }
@@ -300,15 +301,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：欠费核销金额| |参数的约束及描述：欠费核销金额|
      * 
      * @return writeoffAmount */
-    public Double getWriteoffAmount() {
+    public BigDecimal getWriteoffAmount() {
         return writeoffAmount;
     }
 
-    public void setWriteoffAmount(Double writeoffAmount) {
+    public void setWriteoffAmount(BigDecimal writeoffAmount) {
         this.writeoffAmount = writeoffAmount;
     }
 
-    public BillSumRecordInfoV2 withCashAmount(Double cashAmount) {
+    public BillSumRecordInfoV2 withCashAmount(BigDecimal cashAmount) {
         this.cashAmount = cashAmount;
         return this;
     }
@@ -316,15 +317,15 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：现金账户金额。| |参数的约束及描述：现金账户金额。|
      * 
      * @return cashAmount */
-    public Double getCashAmount() {
+    public BigDecimal getCashAmount() {
         return cashAmount;
     }
 
-    public void setCashAmount(Double cashAmount) {
+    public void setCashAmount(BigDecimal cashAmount) {
         this.cashAmount = cashAmount;
     }
 
-    public BillSumRecordInfoV2 withCreditAmount(Double creditAmount) {
+    public BillSumRecordInfoV2 withCreditAmount(BigDecimal creditAmount) {
         this.creditAmount = creditAmount;
         return this;
     }
@@ -332,11 +333,11 @@ public class BillSumRecordInfoV2 {
     /** |参数名称：信用账户金额。| |参数的约束及描述：信用账户金额。|
      * 
      * @return creditAmount */
-    public Double getCreditAmount() {
+    public BigDecimal getCreditAmount() {
         return creditAmount;
     }
 
-    public void setCreditAmount(Double creditAmount) {
+    public void setCreditAmount(BigDecimal creditAmount) {
         this.creditAmount = creditAmount;
     }
