+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/DemandDiscountRatingResult.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /** DemandDiscountRatingResult */
@@ -21,7 +22,7 @@ public class DemandDiscountRatingResult {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -65,7 +66,7 @@ public class DemandDiscountRatingResult {
         this.discountType = discountType;
     }
 
-    public DemandDiscountRatingResult withAmount(Double amount) {
+    public DemandDiscountRatingResult withAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
@@ -73,11 +74,11 @@ public class DemandDiscountRatingResult {
     /** 优惠金额
      * 
      * @return amount */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
