+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/DemandProductRatingResult.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -24,17 +25,17 @@ public class DemandProductRatingResult {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "discount_amount")
 
-    private Double discountAmount;
+    private BigDecimal discountAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "official_website_amount")
 
-    private Double officialWebsiteAmount;
+    private BigDecimal officialWebsiteAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -78,7 +79,7 @@ public class DemandProductRatingResult {
         this.productId = productId;
     }
 
-    public DemandProductRatingResult withAmount(Double amount) {
+    public DemandProductRatingResult withAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
@@ -86,15 +87,15 @@ public class DemandProductRatingResult {
     /** |参数名称：总额| |参数约束及描述：即最终优惠的金额|
      * 
      * @return amount */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
-    public DemandProductRatingResult withDiscountAmount(Double discountAmount) {
+    public DemandProductRatingResult withDiscountAmount(BigDecimal discountAmount) {
         this.discountAmount = discountAmount;
         return this;
     }
@@ -102,15 +103,15 @@ public class DemandProductRatingResult {
     /** |参数名称：优惠额（官网价和总价的差）| |参数约束及描述：优惠额（官网价和总价的差）|
      * 
      * @return discountAmount */
-    public Double getDiscountAmount() {
+    public BigDecimal getDiscountAmount() {
         return discountAmount;
     }
 
-    public void setDiscountAmount(Double discountAmount) {
+    public void setDiscountAmount(BigDecimal discountAmount) {
         this.discountAmount = discountAmount;
     }
 
-    public DemandProductRatingResult withOfficialWebsiteAmount(Double officialWebsiteAmount) {
+    public DemandProductRatingResult withOfficialWebsiteAmount(BigDecimal officialWebsiteAmount) {
         this.officialWebsiteAmount = officialWebsiteAmount;
         return this;
     }
@@ -118,11 +119,11 @@ public class DemandProductRatingResult {
     /** |参数名称：官网价| |参数约束及描述：官网价|
      * 
      * @return officialWebsiteAmount */
-    public Double getOfficialWebsiteAmount() {
+    public BigDecimal getOfficialWebsiteAmount() {
         return officialWebsiteAmount;
     }
 
-    public void setOfficialWebsiteAmount(Double officialWebsiteAmount) {
+    public void setOfficialWebsiteAmount(BigDecimal officialWebsiteAmount) {
         this.officialWebsiteAmount = officialWebsiteAmount;
     }
