+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/FreeResourceDetail.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /** FreeResourceDetail */
@@ -46,12 +47,12 @@ public class FreeResourceDetail {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "original_amount")
 
-    private Double originalAmount;
+    private BigDecimal originalAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -170,7 +171,7 @@ public class FreeResourceDetail {
         this.endTime = endTime;
     }
 
-    public FreeResourceDetail withAmount(Double amount) {
+    public FreeResourceDetail withAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
@@ -178,15 +179,15 @@ public class FreeResourceDetail {
     /** |参数名称：免费资源剩余额度| |参数的约束及描述：免费资源剩余额度|
      * 
      * @return amount */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
-    public FreeResourceDetail withOriginalAmount(Double originalAmount) {
+    public FreeResourceDetail withOriginalAmount(BigDecimal originalAmount) {
         this.originalAmount = originalAmount;
         return this;
     }
@@ -194,11 +195,11 @@ public class FreeResourceDetail {
     /** |参数名称：免费资源原始额度| |参数的约束及描述：免费资源原始额度|
      * 
      * @return originalAmount */
-    public Double getOriginalAmount() {
+    public BigDecimal getOriginalAmount() {
         return originalAmount;
     }
 
-    public void setOriginalAmount(Double originalAmount) {
+    public void setOriginalAmount(BigDecimal originalAmount) {
         this.originalAmount = originalAmount;
     }
