+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/ListPostpaidBillSumResponse.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.huaweicloud.sdk.core.SdkResponse;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /** Response Object */
@@ -17,27 +18,27 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "initial_amount_due")
 
-    private Double initialAmountDue;
+    private BigDecimal initialAmountDue;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "consume_amount")
 
-    private Double consumeAmount;
+    private BigDecimal consumeAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "refunds")
 
-    private Double refunds;
+    private BigDecimal refunds;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "adjustments")
 
-    private Double adjustments;
+    private BigDecimal adjustments;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "tax_amount")
 
-    private Double taxAmount;
+    private BigDecimal taxAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "currency")
@@ -60,7 +61,7 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
         this.billCycle = billCycle;
     }
 
-    public ListPostpaidBillSumResponse withInitialAmountDue(Double initialAmountDue) {
+    public ListPostpaidBillSumResponse withInitialAmountDue(BigDecimal initialAmountDue) {
         this.initialAmountDue = initialAmountDue;
         return this;
     }
@@ -68,15 +69,15 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     /** |参数名称：初始应还金额（含税）| |参数的约束及描述：初始应还金额（含税）|
      * 
      * @return initialAmountDue */
-    public Double getInitialAmountDue() {
+    public BigDecimal getInitialAmountDue() {
         return initialAmountDue;
     }
 
-    public void setInitialAmountDue(Double initialAmountDue) {
+    public void setInitialAmountDue(BigDecimal initialAmountDue) {
         this.initialAmountDue = initialAmountDue;
     }
 
-    public ListPostpaidBillSumResponse withConsumeAmount(Double consumeAmount) {
+    public ListPostpaidBillSumResponse withConsumeAmount(BigDecimal consumeAmount) {
         this.consumeAmount = consumeAmount;
         return this;
     }
@@ -84,15 +85,15 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     /** |参数名称：消费金额| |参数的约束及描述：消费金额|
      * 
      * @return consumeAmount */
-    public Double getConsumeAmount() {
+    public BigDecimal getConsumeAmount() {
         return consumeAmount;
     }
 
-    public void setConsumeAmount(Double consumeAmount) {
+    public void setConsumeAmount(BigDecimal consumeAmount) {
         this.consumeAmount = consumeAmount;
     }
 
-    public ListPostpaidBillSumResponse withRefunds(Double refunds) {
+    public ListPostpaidBillSumResponse withRefunds(BigDecimal refunds) {
         this.refunds = refunds;
         return this;
     }
@@ -100,15 +101,15 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     /** |参数名称：退款| |参数的约束及描述：退款|
      * 
      * @return refunds */
-    public Double getRefunds() {
+    public BigDecimal getRefunds() {
         return refunds;
     }
 
-    public void setRefunds(Double refunds) {
+    public void setRefunds(BigDecimal refunds) {
         this.refunds = refunds;
     }
 
-    public ListPostpaidBillSumResponse withAdjustments(Double adjustments) {
+    public ListPostpaidBillSumResponse withAdjustments(BigDecimal adjustments) {
         this.adjustments = adjustments;
         return this;
     }
@@ -116,15 +117,15 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     /** |参数名称：调账| |参数的约束及描述：调账|
      * 
      * @return adjustments */
-    public Double getAdjustments() {
+    public BigDecimal getAdjustments() {
         return adjustments;
     }
 
-    public void setAdjustments(Double adjustments) {
+    public void setAdjustments(BigDecimal adjustments) {
         this.adjustments = adjustments;
     }
 
-    public ListPostpaidBillSumResponse withTaxAmount(Double taxAmount) {
+    public ListPostpaidBillSumResponse withTaxAmount(BigDecimal taxAmount) {
         this.taxAmount = taxAmount;
         return this;
     }
@@ -132,11 +133,11 @@ public class ListPostpaidBillSumResponse extends SdkResponse {
     /** |参数名称：销项税金额| |参数的约束及描述：销项税金额|
      * 
      * @return taxAmount */
-    public Double getTaxAmount() {
+    public BigDecimal getTaxAmount() {
         return taxAmount;
     }
 
-    public void setTaxAmount(Double taxAmount) {
+    public void setTaxAmount(BigDecimal taxAmount) {
         this.taxAmount = taxAmount;
     }
